docs: add a language pack "contract" document (#918)
* docs: restructure & add a language pack "contract"

The docs/guides/langugage-packs.md document does a good job of describing
what's required of a language pack in terms of directory structure and the
options for the `manifest.yaml` file. But it does a pretty crappy job of
describing the whole thing in context. I've tried to do that here.

Also some fixes to links and reorganization of the docs to make things a
little easier to find.

Signed-off-by: Lance Ball <lball@redhat.com>

* fixup: inline language pack referenc into contract

Signed-off-by: Lance Ball <lball@redhat.com>

* fixup: minor wording tweak

Signed-off-by: Lance Ball <lball@redhat.com>

* fixup: whitespace

Signed-off-by: Lance Ball <lball@redhat.com>