fix: add paketo builder by default to func.yaml
Removes the default builder image lookup map (which resulted in the same
returned value for all supported runtimes), and instead uses that as the
default when creating a new project from the default builtin templates.
This means that we will not default to using the paketo builder for
external language packs, but instead rely on the language pack authors
to specify this value explicitly.

This change is necessary to support on cluster builds using Tekton
Pipelines.

Tangentially, this commit also removes the `builders` array from
func.yaml. Now that I am writing this commit message, I am wondering if
that means that I should also write a migration for it. Hmmm... The
array was removed because it is no longer needed/relevant/used.

Also, unrelated to the original intent of this commit is the removal of
the health endpoints from the top level manifest.yaml in the default
templates. These defaults are already handled in repository.go (where a
copy and paste error was also addressed).

Fixes: https://github.com/knative-sandbox/kn-plugin-func/issues/1005

Signed-off-by: Lance Ball <lball@redhat.com>