chore: change Makefile for multiple tags at HEAD
With the move to the full Knative release train, we now get two tags on
HEAD for the release commit. In the current case, it was `v0.34.0` for
the "internal" version, and `knative-1.7.0`. The Makefile attempts to
determine the version for static linking by looking at the tag at HEAD.
With the git command returning two tags, `make` broke. This change takes
the first of the two tags (will always be `knative-x.x.x`) which I think
is less confusing for CLI users than `v0.34.0`.

I've also updated docs with a couple of changes that weren't reflected
there but landed in 1.7.

Finally, I've removed versions.txt as that is no longer used, and
docs/RELEASING.md as our processes have changed.

Signed-off-by: Lance Ball <lball@redhat.com>