[release-1.8] techdebt(buildpacks): test builder image prefix (#1441)
* techdebt(buildpacks): test builder image prefix

Adds a test for the known builder image prefixes, ensuring they are considered trusted. Removes the check for old versions of podman in order to facilitate this test without turning it into an integration test. Podman 3.4 was released in the summer of '21, more than a year ago, and all published versions of RHEL ship with podman 4.2.

/kind techdebt

Signed-off-by: Lance Ball <lball@redhat.com>

* fixup: remove unused code

Signed-off-by: Lance Ball <lball@redhat.com>

* fixup: run ./hack/update-codegen.sh

Signed-off-by: Lance Ball <lball@redhat.com>

* fixup: run ./hack/update-codegen.sh

Signed-off-by: Lance Ball <lball@redhat.com>

Signed-off-by: Lance Ball <lball@redhat.com>
Co-authored-by: Lance Ball <lball@redhat.com>