cleanup(buildpacks): tweak TrustedBuilder() (#1444)
* cleanup(builders): tweak TrustedBuilder()

We have been relying on a closure to validate the builder prefix provided to `pack`. This change ensures that we use the builder image which `pack` is using - whether it's what we provided in options or not.

/kind cleanup

Signed-off-by: Lance Ball <lball@redhat.com>

* fixup

Signed-off-by: Lance Ball <lball@redhat.com>

* techdebt(builders-tests): simplify TrustedBuilders() unit test

There's no need to mock anything if we don't create the `TrustedBuilders()` function in `Build()`.

/kind techdebt

Signed-off-by: Lance Ball <lball@redhat.com>

Signed-off-by: Lance Ball <lball@redhat.com>