fix: improve error message for missing registry name (#1510)
When a user attempts to build, run or deploy a function and it has not
been previously built, the `--registry` flag or `FUNC_REGISTRY` environment
variable must be set so that the image name can be calculated. The
current error message, `registry required` is not very helpful, and
requires the user to divine what needs to happen next to fix the
problem. This change updates the error message text to indicate that the
user should set the `--registry` flag or `FUNC_REGISTRY` environment
variable.

Fixes: https://github.com/knative/func/issues/1504

Signed-off-by: Lance Ball <lball@redhat.com>

Signed-off-by: Lance Ball <lball@redhat.com>