fix: run cmd tests depend on func.yaml
Updates the run command's tests to not rely on func.yaml, which is an
implementation detail of the client package, instead using the client
API to create the function.

Adds an optional function manipulation "setup" method to the tests which
allows the test to mutate the function, taking the place logically of
providing an exact func.yaml.