chore: use a custom presubmit_tests.sh script
The Knative test infrastructure runs presubmit tests which are defined
in vendor/knative.dev/hack/presubmit-tests.sh. This has been running in
test and release infrastructure. It is possible to customize the
behavior of this script by defining a handful of functions. The function
definitions provided here ensure that we are running tasks that are
already being run with the default (e.g. verify-codgen.sh) but make use
of our existing build and test targets.

(There also seems to be a stray version of this that I committed a
couple of years ago. That has been removed as it is in the wrong
location for test-infra and was never used anyway.)

Relates to: https://github.com/knative/func/issues/1333

Signed-off-by: Lance Ball <lball@redhat.com>